+++ src/angle/UniformEqual.cc
@@ -18,7 +18,7 @@ namespace detran_angle
 
 UniformEqual::UniformEqual(size_t order, size_t dim)
   : Quadrature(dim,
-               std::pow(2, dim) * order * order,
+               std::pow((float)2, (int)dim) * order * order,
                "UniformEqual")
 {
